Fix that deterministic TVars (possible functions) were  used as ND
Testing for isFunFree not required, since this is done at the analysis stage. If  a variable from a determinsitic branch  is brought into scope, it cannot be a function. Functions are never wrapped in deterministic constructors. 
Also: The annotation is always ND for stuff that is known to contain a function.